modules/cloud_function/function/requirements.txt
functions-framework
cloudevents
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
